+++ tests/testthat/test_ror.R
@@ -45,7 +45,8 @@ test_that("ROR outputs are as expected", {
   expect_true(all(names(a2$params) %in% c("test_hyp",
                                           "eval_period",
                                           "null_ratio",
-                                          "alpha")))
+                                          "alpha",
+                                          "cont_adj")))
   expect_is(a2$params$test_hyp, "character")
   expect_equal(a2$params$null_ratio, 1)
   expect_equal(a2$params$alpha, 0.05)
@@ -152,7 +153,8 @@ test_that("ROR df parameter functions as expected", {
   expect_true(all(names(a3$params) %in% c("test_hyp",
                                           "eval_period",
                                           "null_ratio",
-                                          "alpha")))
+                                          "alpha",
+                                          "cont_adj")))
   expect_is(a3$params$test_hyp, "character")
   expect_equal(a3$params$null_ratio, 1)
   expect_equal(a3$params$alpha, 0.05)